requirements.txt
attrs==19.3.0
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.2
python-dateutil==2.8.1
six==1.15.0
wcwidth==0.1.9
